Introduce ComposePermission class for permission management
This commit introduces the ComposePermission class in 'compose-permissions' module. The class uses the newly added PermissionStatus enumeration and it handles permission request processing in the app. The class also provides functions to check the status of the user-given permission, whether it is Granted or Denied. Its main function is to streamline permission requests and checks using the Jetpack Compose paradigm.