Relocate AndroidManifest.xml to new 'example' directory
The AndroidManifest.xml file has been moved from the 'app/src/main' directory to the 'example/src/main' directory. This provides a more organized structure for the project. It also aligns with best practices for Android project organization, making it easier to manage and navigate through the codebase. Packages names have been updated in AndroidManifest and build.gradle files to conform with the new structure.