Add MIT License to the project
A MIT license file has been added to make clear the terms under which the code can be used, copied, modified, merged, published, distributed, sublicensed and/or sold. This is essential for open source projects to attract contributors, assure users of their rights and protections and to avoid legal issues down the line.