Switch repository to GitHub and add Maven publishing
Changed repository from JetBrains Space to GitHub Packages to streamline project management, necessitating the use of GitHub credentials. Also added the maven-publish plugin to enable publishing JAR files to the new repository. A task was created to produce the source JAR from the main source set, and the publications for Maven were outlined in the publishing block, specifying the required parameters and published artifacts.