Update publish settings in build.gradle.kts
Updated the publishing settings in compose-permissions/build.gradle.kts. Changed the Maven publication name from 'maven' to 'release'. Removed the direct reference to the output AAR file and replaced it with a reference to the 'release' component, making it easier to manage the publications. Furthermore, added a block to specify the minimum compile SDK for the AAR metadata and a single variant publish setup with source and javadoc jars. These changes are made to streamline the publishing process and setting up more granular control.