Remove manual dependency management in pom
Removed the Pom section in the 'build.gradle.kts' file for 'compose-permissions' that manually managed dependencies. With the growing project and increased dependencies, using Gradle's in-built dependency management is more efficient and it reduces the chances of errors. It also aligns the project to commonly followed standards, making it easier to onboard new contributors.