Update maven repository URL in Gradle build script
Updated the URL for the Maven repository in the `build.gradle.kts` file for the 'compose-permissions' module. Previously, it was using a placeholder URL, which has been replaced with the actual URL for the package repository on Github (`https://maven.pkg.github.com/jianastrero/compose-permissions`). This allows Gradle to properly fetch the project dependencies during the build process.