Update version and disable minify for compose-permissions
The version of artifactId 'compose-permissions' is upgraded from '1.0.0' to '1.0.1'. Also, in 'buildTypes' the minifyEnabled for 'release' has been disabled and its proguard rules removed. Upgrade in version might be due to minor improvements in the code, and disabling of the minify can help in faster build times and easier debugging.