Add badges to README.md
The badges for Android, Kotlin, Jetpack Compose, and Total Hits Count are added to README.md. These additions aim to provide a quick visual summary of the key technologies used in the library and to indicate its usage frequency; making the library more engaging and adding credibility for the users.