Update README for better understanding and cleaner installation guide
This update simplifies the installation process of the 'Compose Permissions' library in README. The previous Github packages method required two steps and several credentials, the updated version uses Sonatype's OSSRH which only needs a simple dependency in the app's 'build.gradle' file and ensures the library is more easily accessible to a larger audience. Also, a Medium URL is added to provide additional information about the library.