Added rationale support in ComposePermission class
In this commit, the ComposePermission class is updated to support the display of rationale. A new Boolean variable, _shouldShowRationale, is added to check if rationale needs to be shown or not. This update helps users understand why a certain permission is required, enhancing transparency and trust. This is particularly helpful in cases where a user might be unclear why the app is asking for certain permissions.