Update MainActivity.kt, add a snackbar for camera permission rationale
Enhanced camera permission request in ComposePermission class in MainActivity.kt file by adding a snackbar which displays a rationale message. If the camera permission rationale should be shown, a snackbar message is displayed to inform the user to grant permission. If the rationale does not need to be shown, the camera permission is requested directly. This gives a better understanding for the users as to why the permissions are needed and enhances UX transparency.